+++ example/android/app/src/main/kotlin/top/huic/xiao_mi_push_plugin_example/MainActivity.kt
@@ -2,5 +2,4 @@ package top.huic.xiao_mi_push_plugin_example
 
 import io.flutter.embedding.android.FlutterActivity
 
-class MainActivity: FlutterActivity() {
-}
+class MainActivity : FlutterActivity()